- replace AvailableOfflineAction and MakeAvailableOfflineAction with AvailableOfflineAction and a switch
